+++ wasm/src/resize.cpp
@@ -1,4 +1,5 @@
 #include "resize.h"
+#include <cstddef>
 #include <algorithm>
 #include <cmath>
 
@@ -25,12 +26,12 @@ void resize_bilinear(
             const float fx = sx - std::floor(sx);
 
             // Four neighbouring pixels
-            const uint8_t* p00 = src + (y0 * srcWidth + x0) * 4;
-            const uint8_t* p10 = src + (y0 * srcWidth + x1) * 4;
-            const uint8_t* p01 = src + (y1 * srcWidth + x0) * 4;
-            const uint8_t* p11 = src + (y1 * srcWidth + x1) * 4;
+            const uint8_t* p00 = src + (static_cast<size_t>(y0) * srcWidth + x0) * 4;
+            const uint8_t* p10 = src + (static_cast<size_t>(y0) * srcWidth + x1) * 4;
+            const uint8_t* p01 = src + (static_cast<size_t>(y1) * srcWidth + x0) * 4;
+            const uint8_t* p11 = src + (static_cast<size_t>(y1) * srcWidth + x1) * 4;
 
-            uint8_t* out = dst + (dy * dstWidth + dx) * 4;
+            uint8_t* out = dst + (static_cast<size_t>(dy) * dstWidth + dx) * 4;
             for (int c = 0; c < 4; ++c) {
                 const float top    = p00[c] + fx * (p10[c] - p00[c]);
                 const float bottom = p01[c] + fx * (p11[c] - p01[c]);
@@ -54,8 +55,8 @@ void resize_nearest(
         const int sy = std::clamp(static_cast<int>((dy + 0.5f) * yScale), 0, srcHeight - 1);
         for (int dx = 0; dx < dstWidth; ++dx) {
             const int sx = std::clamp(static_cast<int>((dx + 0.5f) * xScale), 0, srcWidth - 1);
-            const uint8_t* srcPx = src + (sy * srcWidth + sx) * 4;
-            uint8_t*       dstPx = dst + (dy * dstWidth + dx) * 4;
+            const uint8_t* srcPx = src + (static_cast<size_t>(sy) * srcWidth + sx) * 4;
+            uint8_t*       dstPx = dst + (static_cast<size_t>(dy) * dstWidth + dx) * 4;
             dstPx[0] = srcPx[0];
             dstPx[1] = srcPx[1];
             dstPx[2] = srcPx[2];
